release 1.0.0: Rails 5.0-6.0, Ruby 2.4-2.7 coverage
- Removes support for Rails older than 5.0.
- Removes support for Ruby older than 2.4.
- Adds test coverage for:
  - Rails 5.0, 5.2, and 6.0
  - Ruby 2.4, 2.5, 2.6, and 2.7
  - DatabaseCleaner 1.7 and 1.8
