+++ .travis.yml.erb
@@ -1,15 +1,19 @@
 language: ruby
+
 rvm:
-  - 1.9.3
-  - 2.0.0
-  - 2.1.3
+  - 2.4
+  - 2.5
+  - 2.6
+  - 2.7
+
 gemfile:
 <% Dir["gemfiles/*gemfile"].each do |file| %>
   - <%= file %>
 <% end %>
+
 matrix:
-  allow_failures:
-  <% Dir["gemfiles/*ar_edge.gemfile"].each do |file| -%>
-    - gemfile: <%= file %>
-  <% end -%>
+  exclude:
+    - rvm: 2.4
+      gemfile: gemfiles/rspec_3.9_ar_6.0_dc_1.8.gemfile
+
 bundler_args: --without development
