+++ Gemfile
@@ -1,13 +1,7 @@
-source "http://rubygems.org"
+source 'http://rubygems.org'
 
 gemspec
 
-group :development do
-  gem 'appraisal', '~> 2.1.0'
-end
-
 group :test do
-  gem 'database_cleaner', '~> 1.3.0'
-  gem 'sqlite3'
-  gem 'rake', '~> 11.0'
+  gem 'rake'
 end
