--- gemfiles/rspec_3.0_ar_4.2.gemfile
@@ -1,18 +1,0 @@
-# This file was generated by Appraisal
-
-source "http://rubygems.org"
-
-gem "activerecord", "~> 4.2.0"
-gem "rspec", "~> 3.0.0"
-
-group :development do
-  gem "appraisal", "~> 2.1.0"
-end
-
-group :test do
-  gem "database_cleaner", "~> 1.3.0"
-  gem "sqlite3"
-  gem "rake", "~> 11.0"
-end
-
-gemspec :path => "../"
