+++ gemfiles/rspec_3.6_ar_5.0.gemfile
@@ -1,0 +1,14 @@
+# This file was generated by Appraisal
+
+source "http://rubygems.org"
+
+gem "activerecord", "~> 5.0.0"
+gem "database_cleaner", "~> 1.7.0"
+gem "rspec", "~> 3.6.0"
+gem "sqlite3", "~> 1.3.13"
+
+group :test do
+  gem "rake"
+end
+
+gemspec path: "../"
