+++ gemfiles/rspec_3.9_ar_5.2_dc_1.7.gemfile
@@ -1,0 +1,14 @@
+# This file was generated by Appraisal
+
+source "http://rubygems.org"
+
+gem "activerecord", "~> 5.2.0"
+gem "database_cleaner", "~> 1.7.0"
+gem "rspec", "~> 3.9.0"
+gem "sqlite3", "~> 1.3.13"
+
+group :test do
+  gem "rake"
+end
+
+gemspec path: "../"
