+++ gemfiles/rspec_3.9_ar_6.0_dc_1.8.gemfile
@@ -1,0 +1,15 @@
+# This file was generated by Appraisal
+
+source "http://rubygems.org"
+
+gem "activerecord", "~> 6.0.0"
+gem "database_cleaner", "~> 1.8.0"
+gem "database_cleaner-active_record", "~> 1.8.0"
+gem "rspec", "~> 3.9.0"
+gem "sqlite3", "~> 1.4.2"
+
+group :test do
+  gem "rake"
+end
+
+gemspec path: "../"
