+++ once-ler.gemspec
@@ -2,19 +2,22 @@
 
 Gem::Specification.new do |s|
   s.name = 'once-ler'
-  s.version = '0.1.4'
+  s.version = '1.0.0'
   s.summary = 'rspec supercharger'
   s.description = "once-ler supercharges your let's and before's with the performance of before(:all)"
 
-  s.required_ruby_version     = '>= 1.9.3'
-  s.required_rubygems_version = '>= 1.3.5'
+  s.required_ruby_version     = '>= 2.4'
+  s.required_rubygems_version = '>= 2.6.0'
 
   s.author            = 'Jon Jensen'
   s.email             = 'jon@instructure.com'
   s.homepage          = 'http://github.com/instructure/once-ler'
 
-  s.files = %w(README.md) + Dir['lib/**/*rb'] + Dir['test/**/*.rb']
-  s.add_dependency('activerecord', '>= 4.0')
-  s.add_dependency('rspec', '>= 3.0')
+  s.files = %w(README.md) + Dir['lib/**/*.rb']
+
+  s.add_dependency 'activerecord', '>= 5.0', '< 6.1'
+  s.add_dependency 'rspec', '>= 3.6'
+
+  s.add_development_dependency 'appraisal', '~> 2.2'
   s.add_development_dependency 'byebug'
 end
